Save time by not prettifying the generated typesLib
The writing of user scripts to consume topics in TypeScript is enhanced
if the data type information from each topic is compiled into TypeScript
  types. All types are compiled into a single file called
  `generatedTypes.ts`, which is a read-only file that is likely never
  opened by the user.

When there is a large amount of topics (and associated data types), the
type generation process can take a long time. Work is ongoing to make
this faster. In the meanwhile, it was discovered that the
`generatedTypes.ts` file is being passed through a prettifier, which can
take upwards of 4-5s during initial loading with 662 topics and 1605
data types in our internal setup.

The resulting code is not a mess either, it looks like the following,
which has somewhat improper indentations and that's about it:

```
    /**
     * MessageTypeBySchemaName enumerates the message types for for all the schema names
     * in the current data source.
     *
     * You probably want to use Message<...> instead.
     */
    export type MessageTypeBySchemaName = {

"action_msgs/GoalInfo": {
"goal_id": MessageTypeBySchemaName["unique_identifier_msgs/UUID"],
"stamp": Time,
},
"unique_identifier_msgs/UUID": {
"uuid": Uint8Array,
},
"action_msgs/GoalStatus": {
 // STATUS_UNKNOWN = 0
 // STATUS_ACCEPTED = 1
 // STATUS_EXECUTING = 2
```