Batch user node state diagnostics and improve layout switching performance
After user nodes are created, the code runs through a loop of
`_setUserNodeDiagnostics` for each user script. These in turn calls
React set state functions. Since set state functions are not batched in
an async function below React 18 (and I see the latest branch is using
React 17), this results in synchronous component updates, which in a
complex layout can result in a 100ms latency per call. In a complex
layout with many scripts (I have 9), this can take about ~2 seconds of
run time (see attached screenshot in PR). Note that I used a previous
build which also had `_setNodeMetadata` in that loop. This means with
the current code and my layout, that would cause a ~1s lag.

Moving this loop into a batched update makes the block only costs 200ms,
see below. With the current code without `setNodeMetadata`, this will
only be 100ms.

Moving to React 18 should remove the need for this.