perf: improve playback performance with the presence of Autocomplete
When there are many `Autocomplete` elements on screen, each of them
needs their width measured. The `Autocomplete` element is being
recreated on forward ref and this is happening for every frame during
playback. Based on flamegraphs, this is taking about 8% CPU time in a
layout with 5 autocomplete elements in a state transition panel, and a
simple plot panel. This is enough to cause choppy rendering as the rest
of the layout is already consistently taking up about 14-16ms of render
time.

This change caches the width in a memo and improves playback performance
such the playback is less choppy.