perf: Improve playback performance with state transition panel
The `StateTransitions` panel creates the chartjs datasets from either
block data ("horizontal" data) or "vertical"/instantaneous data. This
dataset is memo'ed in React based on both the horizontal data
(`decodedBlocks`) and the vertical data (`itemsByPath`).  This memo is
necessary because creating the dataset can be a time consuming operation
(~10-15ms).  Since the vertical data can change every frame, this means
the chartjs datasets are being recreated every frame in the memo.  This
is not a good idea because most of the time, the memo doesn't even use
`itemsByPath` as all data is coming from blocks (as it should? this is a
horizontal panel).  This change is a quick "fix" for the problem where
we do not recompute the datasets if no paths are using the vertical data
and all paths are using the horizontal data. The fix is not ideal, and
the reason is given in the code comments.

Collectively, this is taking ~15-20% CPU time and delays every frame by
10-15ms: