fix seek forward for players that don't implement playUntil
Since the introduction of `TopicAliasingPlayer` between `UserNodePlayer`
and the underlying player, the seek forward button is broken for
underlying players that do not implement the `playUntil` method and
relied on the fallback `seekPlayback` method call. The
`PlaybackControls` and `UserNodePlayer` both has the fallback code paths
for calling `seekPlayback` in case `playUntil` is undefined. This is not
implemented in `TopicAliasingPlayer`. This patch introduces this
fallback code in `TopicAliasingPlayer` to fix this issue.